// types/update-announcement.request.ts
import { AnnouncementStatusEnum } from "./enums/announcement-status.enum";
import { MedicationTypeEnum } from "./enums/medication-type.enum";

export interface UpdateAnnouncementRequest {
  id: string;
  ean: string;
  name: string;
  batch: string;
  expirationDate: string;
  quantity: number;
  status: AnnouncementStatusEnum;
  requiresRefrigeration: boolean;
  medicationType: MedicationTypeEnum | "none";
  activeIngredient: string;
  contentDescription: string;
  classification: string;
  requiresPrescription: boolean;
  administrationRoute: string;
  usageInstructions: string;
  conservation: string;
  idCategory: string;
  price: number;
  basePrice: number;
  dynamicPrice: boolean;
  dynamicPriceUnit: number;
  dynamicPriceUnitValue: number;
  dynamicPricePercent: number;
  dynamicTotalPrice: number;
}